+++ src/hkimjp/jpy/scoreboard.clj
@@ -8,9 +8,10 @@
    [hkimjp.jpy.util :refer [score]]
    [hkimjp.jpy.view :refer [page]]))
 
-(defonce users (mapv :login (-> @(client/get (env :users))
-                                :body
-                                (charred/read-json :key-fn keyword))))
+(defn users []
+  (sort (mapv :login (-> @(client/get (env :users))
+                         :body
+                         (charred/read-json :key-fn keyword)))))
 
 (def smile (constantly "😀"))
 
@@ -21,9 +22,9 @@
   (page
    [:div.m-4
     [:div.text-2xl.font-medium "scoreboard"]
-    (for [user users]
-      [:div.flex.gap-x-4
+    (for [user (users)]
+      [:div.flex.mx-4
        (if (= identity user)
-         [:div {:class "w-24 text-white bg-red-500"} user]
+         [:div {:class "w-24"} [:span.text-white.bg-red-500 user]]
          [:div {:class "w-24"} user])
        [:div (submits user)]])]))