+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 # 0.7.3
 
-- system/users
+- scoreboard/users
 
 # 0.7.2 (2026-03-13)